requirements.txt
backcall==0.2.0
certifi==2022.12.7
charset-normalizer==2.0.12
cycler==0.11.0
dataclasses==0.6
decorator==5.1.1
entrypoints==0.4
filelock==3.4.1
huggingface-hub==0.4.0
idna==3.4
importlib-metadata==4.8.3
importlib-resources==5.4.0
ipykernel==5.5.6
ipython==7.16.3
ipython-genutils==0.2.0
jedi==0.17.2
jupyter-client==7.1.2
jupyter-core==4.9.2
kaggle==1.5.13
kiwisolver==1.3.1
matplotlib==3.3.4
nest-asyncio==1.5.6
numpy==1.19.5
packaging==21.3
pandas==1.1.5
parso==0.7.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.4.0
prompt-toolkit==3.0.36
ptyprocess==0.7.0
Pygments==2.14.0
pyparsing==3.0.9
python-dateutil==2.8.2
python-slugify==6.1.2
pytz==2022.7.1
PyYAML==6.0
pyzmq==25.0.1
requests==2.27.1
six==1.16.0
text-unidecode==1.3
timm==0.6.12
torch==1.9.1+cu111
torchaudio==0.9.1
torchvision==0.10.1+cu111
tornado==6.1
tqdm==4.64.1
traitlets==4.3.3
typing_extensions==4.1.1
urllib3==1.26.15
wcwidth==0.2.6
zipp==3.6.0
